Moved files into folders for better overview.
