+++ Battleship/Display.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Battleship.Utilities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
