+++ Battleship/DisplayElements/StatusBar.cs
@@ -1,0 +1,17 @@
+﻿using Battleship.Utilities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Battleship.DisplayElements
+{
+    public class StatusBar : DisplayElement
+    {
+        public override void Redraw()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
