+++ Battleship/Enums/Orientation.cs
@@ -1,4 +1,4 @@
-﻿namespace Battleship
+﻿namespace Battleship.Enums
 {
     public enum Orientation
     {
