+++ Battleship/Game.cs
@@ -1,3 +1,4 @@
+using Battleship.DisplayElements;
 using System;
 using System.Collections.Generic;
 using System.Linq;
