+++ Battleship/GameModels/Ship.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using Battleship.Enums;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Battleship
+namespace Battleship.GameModels
 {
     public class Ship
     {
